Thu, Nov 3, 2022, 11:11 AM +00:00

+++ filter-less/helpers.c
@@ -1,7 +1,7 @@
 #include "helpers.h"
 
 // Swap two values
-void swap(int *a, int *b)
+void swap(BYTE *a, BYTE *b)
 {
     int tmp = *a;
     *a = *b;
